Improve & document error handling in benchmarks
